flesh out async support in `wit-dylib` (#2367)
* flesh out async support in `wit-dylib`

This does two things:

First, it adds a low-level API to `wit-dylib-ffi` for async imports and exports
as an alternative to the high-level API based on `wit-bindgen`'s `rt` library
and enabled with the `async` feature.  Thus enabling the `async` feature now
really just means "use the high-level API", while disabling it means "use the
low-level API".  The low-level API can make more sense for e.g. Python, where
the coroutine model is different enough from Rust's that it's easier to map it
directly to the component model ABI rather than try to plumb it through Rust's
async APIs.

Second, it populates the `wit_future` and `wit_stream` types with all the
relevant intrinsics, ABI size/align metadata, and generated payload lift and
lower functions.

I've tested this all end-to-end with `componentize-py`, but could use some
guidance on how to add tests here as well.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* split `wit-dylib` `async` feature into `-raw` and `-runtime` variations

- `async-runtime`: integrates with `std::future` by way of `wit_bindgen::rt`
- `async-raw`: minimal layer over raw bindings; does not use `std::future` or `wit_bindgen`

Note that these are mutually exclusive; enabling both will result in a compiler
error.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* remove unnecesary stack frame from `lift_payload`

...and add a comment explaining why it's still needed in `lower_payload`.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* add stream and future tests

Signed-off-by: Joel Dice <joel.dice@fermyon.com>

* Install dev tag of wasmtime

---------

Signed-off-by: Joel Dice <joel.dice@fermyon.com>
Co-authored-by: Alex Crichton <alex@alexcrichton.com>